% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additionalfunctions.R
\name{amse}
\alias{amse}
\title{Function average Mean Square Error 'AMSE'
For a given data f, the function 'AMSE' returns the
average mean square error for the data over all samples}
\usage{
amse(f)
}
\arguments{
\item{f}{data of an n x nx matrix, where n is number of samples (rows) and nx reperesents the grid size}
}
\description{
Function average Mean Square Error 'AMSE'
For a given data f, the function 'AMSE' returns the
average mean square error for the data over all samples
}
