% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additionalfunctions.R
\name{mse}
\alias{mse}
\title{--- AuxFun for adding knots ---#
------------------------#
Function mean squared error 'mse'
For a given vector x, the function 'mse' returns the the mean squared error,
mse measures the average of the squares of the errors}
\usage{
mse(x)
}
\arguments{
\item{x}{data of an n x nx matrix, where n is number of samples (rows) and nx reperesents the grid size}
}
\description{
--- AuxFun for adding knots ---#
------------------------#
Function mean squared error 'mse'
For a given vector x, the function 'mse' returns the the mean squared error,
mse measures the average of the squares of the errors
}
