% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_split.R
\name{opt_split}
\alias{opt_split}
\title{Finding the optimal splitting point which minimize the average mean square error over the range of the data.}
\usage{
opt_split(f, AMSE, M = 5)
}
\arguments{
\item{f}{n x nx matrix. where n is number of samples (rows) and nx reperesents the grid size
in functional data interpreptation, AMSE is the total of mean square errors computed rowwise}

\item{AMSE}{integer.  the average mean square errors of \code{f} computed over its range.}

\item{M}{integer.  the minimal number of points between the optimal split  and the two ends of the iterval.
The default is 10.  The program will return \code{Na} if there are less than \code{2*M+1} points in the range \code{f}.}
}
\value{
A list made of three numeric values:  \code{opt_ix}, \code{AMSE_L} and \code{AMSE_R}.
The first numiric value \code{opt_ix} is the optimal split (knot), and  \code{AMSE_L}, \code{AMSE_R}
are the average mean square errors left and right the spliting point, respectively.
}
\description{
For a given data \code{f} and its average mean square error \code{AMSE}, the function returns a vector of
the optimal split with the average mean square errors on the left and right of the splitting point, respectively.
}
\section{References}{
 Nassar, H., Podgórski, K. (2019) Empirically driven
orthonormal bases for functional data analysis. \emph{Preprint}. Department of Statistics, Lund University.
}

\examples{
## Example:
n=10
f=rbetafda(n)
nx=dim(f)[2]
AMSE = amse(f) ### Total mean square error for whole the samples
opt_split(f,AMSE)
}
\seealso{
\code{split} for constructing split at a given knot;
}
