% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbetafda.R
\name{rbetafda}
\alias{rbetafda}
\title{Function to produce generic functional data}
\usage{
rbetafda(n, nx = 500, ta = 1, tb = 1, sigma = 0, K = 1000, seed)
}
\arguments{
\item{n}{number of samples}

\item{nx}{number of coloumns}

\item{ta}{is shape parameter}

\item{tb}{is shape parameter}

\item{seed}{is a seeding option for reproducibility issues}
}
\value{
The function produces a fda random sample of size n over the equidistant grid of size nx, 1/(nx+1):1/(nx+1):nx/(nx+1)
with equal to square roots of beta densities with parameters alpha and beta simulated independently from gamma distributions with
the shape parameters ta and tb, respectively, and normalized to yield the mean 0.5.
}
\description{
Function to produce generic functional data
}
\examples{
f1=rbetafda(10)
}
