% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{split}
\alias{split}
\title{Evaluating the average mean square errors on the left and right intevals of a given split i}
\usage{
split(f, i)
}
\arguments{
\item{f}{data of an n x nx matrix, where n is number of samples (rows) and nx reperesents the grid size}

\item{i}{an integer withen the range of the object, i.e \eqn{(1 , (nx-1))}, indecates the split.}
}
\value{
vector of two average mean square errors (AMSE_L, AMSE_R):left and righr the spliting point, respectively
}
\description{
For a given data  \code{f} and a given split \code{i}, the function returns vector of
two values of the average mean square errors on the left and righe intevals generated by this split.
}
\section{References}{

Nassar, H., Podgórski, K. (2019) Empirically driven orthonormal bases for functional data analysis. \emph{Preprint}.
Department of Statistics, Lund University.
}

\examples{
n=10
f=rbetafda(n)
nx=dim(f)[2]
xx=vector()
for( i in 1:499){
  Q=split(f,i)
  xx[i]=Q[1]*i/nx+Q[2]*(nx-i)/nx
}
plot(xx)
}
\seealso{
\code{opt_split}
}
