% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wine.R
\docType{data}
\name{Wine}
\alias{Wine}
\title{For the purpose of testing the efficiency of DDK methods we choose the classical functional wine spectra dataset. The wine dataset is provided by
Professor Marc Meurens of the University Catholique de Louvain.}
\format{
A list that contains for arrays x.learning, y.learning, x.test, y.test
}
\source{
\url{https://github.com/cran/cggd/tree/master/data/}
}
\usage{
Wine
}
\description{
The data set consists of results collected from an experiment measuring the alcohol concentration in wine samples by mid infrared spectroscopy.
The training and validation sets contain 94 and 30 spectra, respectively
}
\keyword{datasets}
